refactor: 상속이 아닌 프로토콜 채택으로 구현
POP 글을 읽던 중, 프로젝트에서 사용했던 많은 상속들을 제거할 수 있다는 생각이 들어서 고치게 되었습니다.
